+++ pypeit/tests/test_metadata.py
@@ -55,7 +55,7 @@ def test_read_combid():
 
 @dev_suite_required
 def test_lris_red_multi_400():
-    file_list = glob.glob(os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA', 'Keck_LRIS_red',
+    file_list = glob.glob(os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA', 'keck_lris_red',
                           'multi_400_8500_d560', '*.fits.gz'))
     cfg_lines = ['[rdx]',
                  'spectrograph = keck_lris_red',
@@ -76,7 +76,7 @@ def test_lris_red_multi_400():
 
 @dev_suite_required
 def test_lris_red_multi():
-    file_list = glob.glob(os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA', 'Keck_LRIS_red',
+    file_list = glob.glob(os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA', 'keck_lris_red',
                           'multi*', '*.fits*'))
     cfg_lines = ['[rdx]',
                  'spectrograph = keck_lris_red',
@@ -96,7 +96,7 @@ def test_lris_red_multi():
 @dev_suite_required
 def test_lris_red_multi_run():
     # Perform the setup
-    file_list = glob.glob(os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA', 'Keck_LRIS_red',
+    file_list = glob.glob(os.path.join(os.environ['PYPEIT_DEV'], 'RAW_DATA', 'keck_lris_red',
                           'multi*', '*.fits*'))
     cfg_lines = ['[rdx]',
                  'spectrograph = keck_lris_red',