+++ CHANGES.rst
@@ -1,4 +1,12 @@
 
+0.12.3dev
+---------
+
+- Implemented DataContainer
+- Added fits I/O methods
+- Implemented SlitTraceSet
+- Modification to add_sobj() for numpy 18
+
 0.12.2 (14 Jan 2019)
 --------------------